refactor: board 생성 및 기물 초기 배치 관련 분리
